+++ src/Source/Domain/DigitalWorldOnline.Commons/Models/Base/ItemModel.cs
@@ -31,7 +31,7 @@ namespace DigitalWorldOnline.Commons.Models.Base
         /// <summary>
         /// Tamer shop selling price in bits.
         /// </summary>
-        public int TamerShopSellPrice { get; set; }
+        public long TamerShopSellPrice { get; set; }
 
         /// <summary>
         /// Attribute multiplier.
