+++ src/Source/Infra/DigitalWorldOnline.Infraestructure/Migrations/DatabaseContextModelSnapshot.cs
@@ -2617,7 +2617,7 @@ namespace DigitalWorldOnline.Infraestructure.Migrations
                         .HasColumnType("int");
 
                     b.Property<int>("TamerShopSellPrice")
-                        .HasColumnType("int");
+                        .HasColumnType("bigint");
 
                     b.HasKey("Id");
